Added clarification for nested requests
Added a sub-section in the Requests portion to clarify the usage of nested calls a bit. You did list `posts/1/comments` in the list of routes, but it's kinda left up to the reader to figure out what that means. Hopefully this makes it a bit easier to pick up quickly. References [#72](https://github.com/typicode/json-server/issues/72)