chore: binding host to ::1 (0.0.0.0) when no host parameter is provided
When trying to run from a docker container, binding the listener to 'localhost'
doesn't allow it to be reachable from outside the container. By removing the host
from app.listen method, it makes express to bind the listener to ::1 (0.0.0.0),
accepting requests for any hostname. If a value different than the default is
provided, it binds it accordingly.